fix: Convert window start calculation to float for precision
This commit updates the `calculate_taes_metrics` function in `metrics.py` to ensure that the window start times are calculated as floats. This change enhances numerical precision in the evaluation process while maintaining existing functionality. The adjustment aligns with previous updates that emphasize the use of constants for improved code maintainability and readability.

+++ src/brain_brr/eval/metrics.py
@@ -719,7 +719,7 @@ def calculate_taes_metrics(
     # Create dummy metadata: treat all windows as from single recording
     n_windows = predictions.shape[0]
     file_ids = ["test_recording"] * n_windows
-    window_starts = [i * constants.STRIDE_SIZE_SEC for i in range(n_windows)]
+    window_starts = [float(i * constants.STRIDE_SIZE_SEC) for i in range(n_windows)]
     metrics = evaluate_predictions(
         predictions,
         labels,
